pyproject.toml
[project]
name = "athys-bot"
version = "0.1.0"
description = "A Discord chat bot for my channel"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "google-genai>=1.19.0",
]
